+++ TrackForce/src/main/java/com/revature/services/BatchesService.java
@@ -1,7 +1,6 @@
 package com.revature.services;
 
 import java.io.IOException;
-import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.Collection;
 import java.util.LinkedList;
@@ -77,11 +76,11 @@ public class BatchesService implements Service {
         return batches;
     }
 
-    private Map<BigDecimal, BatchInfo> getBatches() throws IOException {
+    private Map<Integer, BatchInfo> getBatches() throws IOException {
         Session session = sessionFactory.openSession();
         Transaction tx = session.beginTransaction();
         try {
-            Map<BigDecimal, BatchInfo> map = batchDao.getBatchDetails(session);
+            Map<Integer, BatchInfo> map = batchDao.getBatchDetails(session);
 
             session.flush();
             tx.commit();
@@ -180,13 +179,13 @@ public class BatchesService implements Service {
     public Set<AssociateInfo> getAssociates(@PathParam("batch") String batchIdStr) throws IOException {
         Set<AssociateInfo> associatesList = PersistentStorage.getStorage()
                 .getBatchAsMap()
-                .get(new BigDecimal(Integer.parseInt(batchIdStr)))
+                .get(new Integer(Integer.parseInt(batchIdStr)))
                 .getAssociates();
         if (associatesList == null) {
             execute();
             return PersistentStorage.getStorage()
                     .getBatchAsMap()
-                    .get(new BigDecimal(Integer.parseInt(batchIdStr)))
+                    .get(new Integer(Integer.parseInt(batchIdStr)))
                     .getAssociates();
         }
         return associatesList;
